backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.9
python-dotenv>=1.0
djangorestframework-simplejwt==5.5.1
djangorestframework-simplejwt[token_blacklist]==5.5.1
django-filter==23.5
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.1.1